feat:TE7-T130: Action Mailer basics and file upload
Users can upload the profile image
display their profile image in the list of users index page.
Generate a UserMailer in your App.
The user should have a name, email
Create a welcome email template and use it
Send welcome mail (Both HTML and text format) when the user is created.
Send an email when an email address is altered. Use different layouts for the emails.
Sending email configurations for the local machines and sending emails in real time